+++ examples/oauth2/src/web/oauth.rs
@@ -31,7 +31,7 @@ mod get {
     use super::*;
 
     pub async fn callback(
-        mut auth_session: AuthSession,
+        auth_session: AuthSession,
         session: Session,
         Query(AuthzResp {
             code,